Improve context manager implementation
You should call the 'connect' method of the 'SSHClient' object in the
'__enter__' function. As you have it, you risk a leak where an
SshConnection object in initialized but never used in a context manager;
when it is deleted without ever being used in a context manager,
'__exit__' will never be called, and the open connection won't be
closed.